Add cell filter actions to data grid component
